Move to package root
